broke up query code
